frontend/requirements.txt
# Requirements for Multimodal Tweet Classification Streamlit App

# Core ML libraries
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0

# Streamlit and UI
streamlit>=1.25.0

# Image processing
Pillow>=8.3.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0

# Additional ML utilities
scikit-learn>=1.0.0
tqdm>=4.62.0

# Optional: For better performance
# accelerate>=0.20.0  # For faster transformer loading